requirements.txt
streamlit
pandas
altair
google-cloud-aiplatform
google-auth
google-auth-oauthlib
google-auth-httplib2
